[autonomous] refactor: split vault route into validation/auth/access/crypto modules (CF-024a)
routes/vault.ts becomes the orchestrator; validateVaultRoute and
handleVaultRequest stay exported for src/index.ts and vault-html.ts.
